// packages/nancalui-vue/nancalui/date-picker-pro/src/date-picker-pro-types.ts
export const datePickerProCommonProps = {
  format: {
    type: String,
  },
  valueFormat: {
    type: String,
  },
  showTime: {
    type: Boolean,
    default: false,
  },
  size: {
    type: String as PropType<InputSize>,
    default: 'md',
  },
  disabled: {
    type: Boolean,
    default: false,
  },
  calendarRange: {
    type: Array as PropType<number[]>,
    default: [1970, 2099],
  },
  limitDateRange: {
    type: Array as PropType<Date[]>,
  },
  type: {
    type: String,
    default: 'date',
  },
};

export const datePickerProProps = {
  modelValue: {
    type: [Date, String] as PropType<Date | string>,
    default: '',
  },
  placeholder: {
    type: String,
    default: '',
  },
  ...datePickerProCommonProps,
} as const;

export type DatePickerProProps = ExtractPropTypes<typeof datePickerProProps>;

export interface UseDatePickerProReturnType {
  originRef: Ref<HTMLElement | undefined>;
  inputRef: Ref<HTMLElement | undefined>;
  overlayRef: Ref<HTMLElement | undefined>;
  isPanelShow: Ref<boolean>;
  placeholder: ComputedRef<string>;
  format: ComputedRef<string>;
  valueFormat: ComputedRef<string>;
  dateValue: ComputedRef<Dayjs | undefined>;
  displayDateValue: ComputedRef<string>;
  isMouseEnter: Ref<boolean>;
  showCloseIcon: ComputedRef<boolean>;
  pickerDisabled: ComputedRef<boolean>;
  pickerSize: ComputedRef<'sm' | 'md' | 'lg'>;
  isValidateError: ComputedRef<boolean>;
  onFocus: (e: MouseEvent) => void;
  onSelectedDate: (date: Dayjs, isConfirm?: boolean) => void;
  handlerClearTime: (e: MouseEvent) => void;
}